Move functions into mutate_aks.go
